Please elaborate the following "The Blob service copies blobs on a best-effort basis."
We got the master article changed recently with more details. 
"https://learn.microsoft.com/en-us/rest/api/storageservices/copy-blob?tabs=azure-ad#remarks" 

Earlier the article said "The Blob service copies blobs on a best-effort basis" but now it has been updated to say "Due to the asynchronous nature of the copy operation, Blob Storage copies blobs on a best-effort basis. The Blob service copies blobs when server resources are not being utilized by other tasks, so a copy is not guaranteed to start immediately or complete in a specified timeframe." 

It would be great to update this article to be in sync with the other one.